kubernetes: adjust AssembleBuild tests
IDs have _ replaced with - in types library, and standardizing allows for
more reuse of the test objects.
I also had to correct the container number for the services.
And I had to add an accurate v1.Pod representation for the _stages pod
now that we're looping over containers to setup environment.